Create  rnn.ipynb
This project was done on google colab. The stock market needs to leverage on the rise of artificial intelligence  in trying to forecast future stock prices. We use the RNN  which can leverage on recalling timestamps of the previous time (n-1) . I  initialized the Re-ccurent Neural Network , then added the layers to kick off the training. The model was tested with new results  and visualization plots were drawn.